Add SystemDiskInterface and NullDiskInterface classes.
Add an intermediate DiskInterface derived class named
SystemDiskInterface that performs real disk i/o without
any intermediate caching.

RealDiskInterface is now a derived class of
SystemDiskInterface that adds a caching layer on
Win32 only. Keeping the same name reduces the number
of changes in this commit to the strict minimum.

NullDiskInterface is provided to simplify custom
DiskInterface implementations used in tests, in
particular because new DiskInterface methods are
going to be added in future commits, and adding
the same method overrides NullDiskInterface will
lower the changes to do in those implementations.